backend_merge/requirements.txt
Flask
Flask-SQLAlchemy
twilio
python-dotenv
web3 # optional, only when implementing chain
